backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
firebase-admin==6.4.0
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0 
googlemaps>=4.10.0
